Explicitly list-convert map before concat w/ list
Under Python 3, trying to concat a map with a list will fail. So this
change updates an existing function in the code that had been doing a
concat of a list and map without first explicitly converting the map to
a list. (That worked under Python 2 but doesn’t under Python 3).